requirements.txt
backports.tempfile==1.0
django==1.11.17
django-allauth==0.38.0
django-mptt==0.9.1
djangorestframework==3.6.4
fabric3==1.14.post1
git+https://github.com/BinaryCrate/behave-django.git@50450684840d66d34c3bd121aa5b75b46f6c22e1
git+https://github.com/historygraphio/historygraph.git@bc568d42f053f5547bbaba3432ebb027efb6bca6
mock==2.0.0
psutil==5.4.8
psycopg2-binary==2.7.6.1
pytest-django==3.4.4

#Wagtail
wagtail==1.13.4

# General requirements for testing
selenium==3.141.0
factory-boy==2.11.1
# Munch is only used in the front end tests
git+https://github.com/BinaryCrate/munch.git@2e302bc9590b7e399b1c4e021e847fb38a1577ec
